requirements/dev.txt
attrs==20.3.0
autopep8==1.5.6
click==7.1.2
Flask==1.1.2
greenlet==1.0.0
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pycodestyle==2.7.0
PyJWT==2.0.1
pyparsing==2.4.7
pytest==6.2.2
python-dotenv==0.15.0
SQLAlchemy==1.4.2
toml==0.10.2
Werkzeug==1.0.1
